dahdi_echocan_mg2: rename ABS define to PARA_ABS
Rename ABS define to PARA_ABS to fix compilation error with mips 24kc
arch.

This target define in his ASM header a similar define ABS that cause
conflicts and redefinition error for this driver.

Rename ABS to PARA_ABS to better suite the usage as it's abs usage with
value decremented by one if it's the max negative int.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>